requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
pymongo
langchain
langchain-core
langchain-google-genai